+++ src/sklearndf/_sklearndf.py
@@ -1,10 +1,10 @@
 """
 Core implementation of :mod:`sklearndf`
 """
-
+import inspect
 import logging
 from abc import ABCMeta, abstractmethod
-from typing import Any, List, Mapping, Optional, Sequence, TypeVar, Union, cast
+from typing import Any, Dict, List, Mapping, Optional, Sequence, TypeVar, Union, cast
 
 import pandas as pd
 from sklearn.base import (
@@ -14,13 +14,23 @@ from sklearn.base import (
     TransformerMixin,
     clone,
 )
+from sklearn.utils import is_scalar_nan
 
-from pytools.api import AllTracker
+from pytools.api import AllTracker, inheritdoc
+from pytools.expression import Expression, HasExpressionRepr, make_expression
+from pytools.expression.atomic import Id
 from pytools.fit import FittableMixin
 
 log = logging.getLogger(__name__)
 
-__all__ = ["EstimatorDF", "LearnerDF", "ClassifierDF", "RegressorDF", "TransformerDF"]
+__all__ = [
+    "EstimatorDF",
+    "LearnerDF",
+    "ClassifierDF",
+    "RegressorDF",
+    "SupervisedLearnerDF",
+    "TransformerDF",
+]
 
 #
 # type variables
@@ -28,6 +38,13 @@ __all__ = ["EstimatorDF", "LearnerDF", "ClassifierDF", "RegressorDF", "Transform
 
 T_Self = TypeVar("T_Self")
 T_EstimatorDF = TypeVar("T_EstimatorDF")
+
+
+#
+# Constants
+#
+
+UNDEFINED = object()
 
 
 #
@@ -42,7 +59,10 @@ __tracker = AllTracker(globals())
 #
 
 
-class EstimatorDF(BaseEstimator, FittableMixin[pd.DataFrame], metaclass=ABCMeta):
+@inheritdoc(match="""[see superclass]""")
+class EstimatorDF(
+    HasExpressionRepr, BaseEstimator, FittableMixin[pd.DataFrame], metaclass=ABCMeta
+):
     """
     Base class for augmented scikit-learn `estimators`.
 
@@ -148,6 +168,71 @@ class EstimatorDF(BaseEstimator, FittableMixin[pd.DataFrame], metaclass=ABCMeta)
         # get the number of outputs this estimator has been fitted to
         pass
 
+    def to_expression(self) -> Expression:
+        """[see superclass]"""
+
+        # create a dictionary with all parameters of the estimator, mapping them
+        # to their default values if defined (and otherwise to Signature.empty)
+        estimator_parameters = {
+            name: parameter.default
+            for name, parameter in inspect.signature(self.__init__).parameters.items()
+        }
+
+        def _kwarg_to_expression(name: str, value: Any) -> Optional[Expression]:
+            expression = make_expression(value)
+
+            # are we dealing with a default value?
+            default_value = estimator_parameters.get(name, UNDEFINED)
+
+            if (
+                (  # there is a parameter with the given name
+                    default_value is not UNDEFINED
+                )
+                and (  # the parameter has a default value
+                    default_value != inspect.Signature.empty
+                )
+                and (
+                    # if the value is an estimator ...
+                    not isinstance(value, BaseEstimator)
+                    # ... then it has the same class as the default value
+                    or value.__class__ == default_value.__class__
+                )
+                and (
+                    # both value and default value are np.nan ...
+                    (is_scalar_nan(value) and is_scalar_nan(default_value))
+                    or (
+                        # ... or both have the same expression.
+                        # We cannot compare for equality since we don't know
+                        # if the classes of the values implement this.
+                        # Therefore we compare the expressions but do this last,
+                        # as it might be computationally more costly in the
+                        # unlikely case that the default value is a very complex
+                        # object.
+                        expression.eq_(make_expression(default_value))
+                    )
+                )
+            ):
+                # we can confirm the given value is the default value - no need
+                # to include in the expression representation of this estimator
+                return None
+
+            else:
+                # custom value: we return the expression
+                return expression
+
+        kwarg_expressions: Dict[str, Optional[Expression]] = {
+            name: _kwarg_to_expression(name, value)
+            for name, value in self.get_params(deep=False).items()
+        }
+
+        return Id(type(self))(
+            **{
+                name: expression
+                for name, expression in kwarg_expressions.items()
+                if expression is not None
+            }
+        )
+
 
 class LearnerDF(EstimatorDF, metaclass=ABCMeta):
     """
@@ -193,6 +278,14 @@ class LearnerDF(EstimatorDF, metaclass=ABCMeta):
             of multiple outputs
         """
         pass
+
+
+class SupervisedLearnerDF(LearnerDF, metaclass=ABCMeta):
+    """
+    Base class for augmented scikit-learn `supervised learners`.
+
+    Provides enhanced support for data frames.
+    """
 
     # noinspection PyPep8Naming
     @abstractmethod
@@ -323,7 +416,7 @@ class TransformerDF(EstimatorDF, TransformerMixin, metaclass=ABCMeta):
         return self.feature_names_original_.index
 
 
-class RegressorDF(LearnerDF, RegressorMixin, metaclass=ABCMeta):
+class RegressorDF(SupervisedLearnerDF, RegressorMixin, metaclass=ABCMeta):
     """
     Base class for augmented scikit-learn `regressors`.
 
@@ -331,7 +424,7 @@ class RegressorDF(LearnerDF, RegressorMixin, metaclass=ABCMeta):
     """
 
 
-class ClassifierDF(LearnerDF, ClassifierMixin, metaclass=ABCMeta):
+class ClassifierDF(SupervisedLearnerDF, ClassifierMixin, metaclass=ABCMeta):
     """
     Base class for augmented scikit-learn `classifiers`.